pyproject.toml
[project]
name = "shacs_bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "croniter>=6.0.0",
    "litellm>=1.81.8",
    "loguru>=0.7.3",
    "mcp>=1.26.0",
    "oauth-cli-kit>=0.1.3",
    "pathlib>=1.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "readability-lxml>=0.8.4.1",
    "rich>=14.3.3",
]
